+++ serious/dict/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ['DictModel']
 
-from .api import DictModel
+from .model import DictModel
