+++ serious/json/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ['JsonModel']
 
-from .api import JsonModel
+from .model import JsonModel
