+++ serious/json/model.py
@@ -5,8 +5,7 @@ from typing import Optional, TypeVar, Type, Generic, List, MutableMapping, Colle
 
 from serious.descriptors import describe
 from serious.preconditions import check_is_instance
-from serious.serialization import FieldSerializer, SeriousModel, field_serializers
-from serious.serialization.model import KeyMapper
+from serious.serialization import FieldSerializer, SeriousModel, field_serializers, KeyMapper
 from serious.utils import class_path, snake_to_camel, camel_to_snake
 from .preconditions import check_that_loading_an_object, check_that_loading_a_list
